model_serving/requirements.txt
keras
optuna
python-box
pyYAML
pyarrow
numpy
pandas
joblib
scikit-learn
ensure
python-dotenv
google-cloud-aiplatform
google-auth
vertexai